Downgraded the clippy:unwrap_in_result lint
Changed the clippy::unwrap_in_result lint from forbid to deny, so that
it can be overridden, as there are certain use cases where it is
acceptable to use unwrap() in a function that returns Result. For
instance, when the outcome is known to be infallible and/or there would
not be a benefit from creating a new error variant specifically for the
failure case.